+++ WrightTools/artists/_interact.py
@@ -203,6 +203,31 @@ def interact2D(
         uniform grids.  Default is False.
     verbose : boolean (optional)
         Toggle talkback. Default is True.
+
+    Returns
+    -------
+    out : types.SimpleNamespace
+        container for important interactive elements of the plot.
+
+        Properties
+        ----------
+        image : 2D artist object
+            The artist of the 2D plot.  Type is either matplotlib.collections.QuadMesh (use_imshow=False) or matplotlib.image.AxesImage (use_imshow=True)
+        sliders : dict (key[name] = value[matplotlib.Widgets.Slider])
+            The sliders.
+        crosshairs: list of horizontal and vertical crosshair
+            crosshairs are lists of matplotlib.lines.Line2D
+        radio : matplotlib.Widgets.RadioButtons
+            radio button for local/global normalization
+        colorbar : matplotlib.colorbar.ColorbarBase
+            The 2D colorbar object
+
+    Usage
+    -----
+    When calling, always store the returned arguments to a variable:
+    >>> out = WrightTools.artists.interact2D(...)
+
+    This prevents interactive buttons from turning off due to python's garbage collection.
     """
     # avoid changing passed data object
     data = data.copy()
@@ -217,7 +242,9 @@ def interact2D(
     if nsliders < 0:
         raise DimensionalityError(">= 2", data.ndim)
     # TODO: implement aspect; doesn't work currently because of our incorporation of colorbar
-    fig, gs = create_figure(width="single", nrows=7 + nsliders, cols=[1, 1, 1, 1, 1, "cbar"])
+    fig, gs = create_figure(
+        width="single", margin=[0.2, 1, 0.2, 1], nrows=7 + nsliders, cols=[1, 1, 1, 1, 1, "cbar"]
+    )
     plt.get_current_fig_manager().set_window_title(f"interact2D: {data.natural_name}")
     # create axes
     ax0 = plt.subplot(gs[1:6, 0:5])
@@ -556,4 +583,11 @@ def interact2D(
     for slider in sliders.values():
         slider.on_changed(update_slider)
 
-    return obj2D, sliders, crosshair_hline, crosshair_vline, radio, colorbar
+    out = SimpleNamespace(
+        image=obj2D,
+        sliders=sliders,
+        crosshairs=[crosshair_hline, crosshair_vline],
+        radio=radio,
+        colorbar=colorbar,
+    )
+    return out