+++ CHANGELOG.md
@@ -8,6 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Added
 - `Data.ichop`, an interation-based version of `Data.chop`
 
+### Fixed
+- fixed Quick2D/Quick1D issues where collapsing unused dims did not work
+- wt5 explore : fixed bug where data will not load interactively if directory is not cwd
+- constants in chopped data will inherit the units of the original data
+
+## Changed
+- artists now gets turbo colormap straight from matplotlib
+- artists.interact2D now returns a `types.SimpleNamespace` object (rather than a tuple)
+
 ## [3.5.2]
 
 ### Added